requirements.txt
anyio==4.2.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-amazon-paapi==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tornado==6.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
